Rename scabbard::store::lmdb to transact
This change renames the module for the LMDB implementation to transact,
in an analogous manner to diesel.
