+++ services/scabbard/libscabbard/src/store/mod.rs
@@ -17,8 +17,8 @@
 #[cfg(feature = "diesel")]
 pub mod diesel;
 mod error;
-#[cfg(feature = "lmdb")]
-pub mod lmdb;
+#[cfg(any(feature = "lmdb", feature = "transact"))]
+pub mod transact;
 
 pub use error::CommitHashStoreError;
